[Close #226] Moved Edit action into namespace.
And another unbuilt dependency, that might have been a Gem, bites the
dust.

RSpec: 696 examples, 0 failures
RuboCop: 237 files inspected, no offenses detected
