--- lib/newpoc-action-post-edit/lib/newpoc/action/post/edit/version.rb
@@ -1,11 +1,0 @@
-
-module Newpoc
-  module Action
-    module Post
-      # Verifies that the current user is permitted to edit a specified post.
-      class Edit
-        VERSION = '0.0.1'
-      end
-    end
-  end
-end
