+++ spec/controllers/posts_controller_spec.rb
@@ -231,7 +231,7 @@ describe PostsController do
     end # context 'for the Guest User'
   end # describe "POST 'create'" (StoreResult removed)
 
-  describe "GET 'edit'" do
+  fdescribe "GET 'edit'" do
     let(:author) { FactoryGirl.create :user, :saved_user }
     let!(:post) do
       FactoryGirl.create :post, :saved_post, :published_post,
